(PC-24812) fix(sonar): move extractStockDates because is specific to offer
